Revamped Movement Script and Obstacles added
Revamped the player movement script to allow for smoother movements using the 'a' and 'd' inputs from the player (velocity based force change instead of default force change). There is a constant forward force now. Also added obstacles in the player's path. Upon hitting one of the obstacles, a message is printed to the console and the player's movement stops. Finally, added fog to limit player's visibiliy of future obstacles